Don't deep copy the manifest unless we're in single threaded mode
Move environment variable handling into dbt.flags so it's easier to check
